% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{AUCfromROCtable}
\alias{AUCfromROCtable}
\title{Compute the area under the curve}
\usage{
AUCfromROCtable(ROC.table)
}
\arguments{
\item{ROC.table}{A result from ROC.table}

\item{...}{Additional parameters are passed to XXXX().}
}
\description{
Compute the area under the curve from a table returned by ROCstatistics
}
\details{
First version: 2015-04
Last modification: 2015-04
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

