% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanEqualityTests.R
\name{meanEqualityTests.compareSignifFeatures}
\alias{meanEqualityTests.compareSignifFeatures}
\title{Return a contingency table for several tests with a meanEqualityTests result}
\usage{
meanEqualityTests.compareSignifFeatures(mnEqlT.result, selected.tests = NULL,
  signif.score = "fdr", alpha = 0.05, plot.venn = FALSE,
  plot.tree = FALSE, hclust.method = "average", plot.heatmap = FALSE, ...)
}
\arguments{
\item{mnEqlT.result}{A result of stats4bioinfo::meanEqualityTests()}

\item{...}{Additional arguments are passed to plot().}

\item{selected.tests=NULL}{Tests to compare. If NULL, all available tests are compared.}

\item{signif.score="fdr"}{Score on which the alpha threshold should be applied (Supported: fdr, p.value, e.value)}

\item{alpha=0.01}{Significance threshold (will be applied on corrected p-values)}

\item{plot.venn=FALSE}{Plot a Venn diagram with the result.}

\item{plot.tree=FALSE}{Plot a hierarchical tree indicating relationships between tests.}

\item{hclust.method="average"}{Hierarchical clustering method (passed to hclust).}

\item{plot.heatmap=FALSE}{Plot a heatmap (in gray scale) highlighting the similarity between feature lists.}
}
\value{
A list containing the following fields:
\describe{
  \item{selected.per.test}{A dataframe with Boolean values indicating the status
  (significant or not) for each row (feature) of the input dataset.}
  \item{contingency}{A contingency table indicating the number of features (rows)
  passing the significance threshold for each pair of tests. The diagonal of this
  contingency table indicates the number of significant features for each individual
  test.}
}
}
\description{
Compare the significant features returned by two or more tests in
an object returned by meanEqualityTest(), and plot the result in the form of a Venn diagram.
}
\details{
First version: 2015-03
Last modification: 2015-03
}
\examples{
## We forward to meanEqualityTests() for examples of utilization
example(meanEqualityTests)

## Compare 2-group mean equality test results with 1\% threshold on FDR
test.compa.2groups <- meanEqualityTests.compareSignifFeatures(diff.results,
                                                    selected.tests=c("student",
                                                                     "welch",
                                                                     "wilcoxon"),
                                                    signif.score="fdr",
                                                    alpha=0.01,
                                                    plot.venn=TRUE)
summary(test.compa.2groups)
print(test.compa.2groups$contingency)

## Compare multi-group mean equality test results with a threshold of 1 on the
## e-value (accept one false positive per analysis)
test.compa.ngroups <- meanEqualityTests.compareSignifFeatures(diff.results,
                                                    selected.tests=c("anova",
                                                                     "kruskal"),
                                                    signif.score="e.value",
                                                    alpha=1)

## Compare all the mean equality test results with a threshold of 1 on the
## e-value (accept one false positive per analysis)
test.compa <- meanEqualityTests.compareSignifFeatures(diff.results,
                                                      selected.tests=c("student",
                                                                       "welch",
                                                                       "wilcoxon",
                                                                       "anova",
                                                                       "kruskal"),
                                                      signif.score="e.value",
                                                      alpha=1,
                                                      plot.venn=FALSE,
                                                      plot.tree=FALSE,
                                                      hclust.method="average",
                                                      main="Common significant probesets",
                                                      xlab="Differential method",
                                                      ylab="Jaccard distance",
                                                      plot.heatmap=TRUE)

                                                                                                                                                        )
summary(test.compa)
print(test.compa$contingency)
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

