% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanEqualityTests.compareTwoTests.R
\name{meanEqualityTests.compareTwoTests}
\alias{meanEqualityTests.compareTwoTests}
\title{Compare the feature lists called significant by two tests from meanEqualityTest object.}
\usage{
meanEqualityTests.compareTwoTests(mnEqlT.result, test1, test2,
  signif.score = "fdr", alpha = 0.05, ...)
}
\arguments{
\item{mnEqlT.result}{A result of stats4bioinfo::meanEqualityTests()}

\item{test1}{First test}

\item{test2}{Second test}

\item{signif.score}{Score on which the alpha threshold should be applied (Supported: fdr, p.value, e.value)}

\item{alpha}{Significance threshold (will be applied on corrected p-values)}

\item{...}{Additional parameters are passed to plotPvalCompa()}
}
\description{
Compare the feature lists  called significant by two tests
(e.g. Welch versus Wilcoxon) from an object returned by meanEqualityTest(),
and plot the result.
}
\details{
First version: 2015-03
Last modification: 2015-03
}
\examples{
## We forward to meanEqualityTests() for examples of utilization
example(meanEqualityTests)

## Compare e-values between Welch and Wilcoxon tests
welch.vs.wilcoxon.e.value <- meanEqualityTests.compareTwoTests(diff.results,
                                               test1="welch",
                                               test2="wilcoxon",
                                               signif.score="e.value",
                                               plot.cex=0.5,legend.cex=0.8)

## Compare FDR between Welch and Wilcoxon tests, draw a grayscal plot
welch.vs.wilcoxon.fdr <- meanEqualityTests.compareTwoTests(diff.results,
                                               test1="welch",
                                               test2="wilcoxon",
                                               alpha=0.01,
                                               signif.score="fdr",
                                               plot.colors=FALSE,
                                               plot.cex=0.5,legend.cex=0.8)
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

