% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanEqualityTests.plotVolcano.R
\name{meanEqualityTests.plotVolcano}
\alias{meanEqualityTests.plotVolcano}
\title{Volcano plot from meanEqualityTest object}
\usage{
meanEqualityTests.plotVolcano(mnEqlT.result, test = "welch",
  main = paste(test, "Volcano plot"), alpha = 0.05,
  legend.corner = "bottomright", plot.cex = 0.7, legend.cex = 1,
  plot.colors = TRUE, ...)
}
\arguments{
\item{mnEqlT.result}{A result of stats4bioinfo::meanEqualityTests()}

\item{alpha}{Significance threshold (will be applied on corrected p-values)}

\item{plot.cex}{Point size for the volcano (passed to graphics::plot() function).}

\item{...}{Additional parameters are passed to plot()}

\item{test="welch"}{Statistical test for which the Volcano plot has to be drawn.}

\item{legend.corner="bottomright"}{Position where legend should be placed.}

\item{legend.cex=1}{Font size factor for legend. This allows to choose separate font sizes to draw the plot itself and the legend.}

\item{plot.colors=TRUE}{Use colors for the plot}
}
\description{
Draw a volcano plot for a selected test of a meanEqualityTest result.
}
\details{
First version: 2015-03
Last modification: 2015-03
}
\examples{
## We forward to meanEqualityTests() for examples of utilization
example(meanEqualityTests)

## Draw a Volcano plot for Welch test results
meanEqualityTests.plotVolcano(diff.results, test="welch")

## Draw a Volcano plot for ANOVA results
meanEqualityTests.plotVolcano(diff.results, test="anova")
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

