% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipleTestingCorrections.R
\name{mulitpleTestingCorrections.plotCorrectedVsPval}
\alias{mulitpleTestingCorrections.plotCorrectedVsPval}
\title{Plot corrected p-values versus nominal p-value}
\usage{
mulitpleTestingCorrections.plotCorrectedVsPval(multitest.result,
  main = "Multitesting corrections", xlab = "p-value",
  ylab = "Multi-testing corrected statistics", alpha = 0.05,
  legend.corner = "topleft", legend.cex = 1, plot.legend = TRUE,
  plot.pch = c(p.value = 2, fdr = 4, qval.0 = 3, e.value = 1, fwer = 20),
  plot.col = c(p.value = "#000000", fdr = "#888888", qval.0 = "#666666",
  e.value = "#BBBBBB", fwer = "#444444"), plot.elements = c("p.value", "fdr",
  "qval.0", "e.value", "fwer"), ...)
}
\arguments{
\item{multitest.result}{the list returned by the function multipleTestingCorrections().}

\item{...}{Additional parameters are passed to plot()}

\item{main='Multitesting}{corrections'  main title of the plot}

\item{alpha=0.05}{Threshold of significance (alpha).}

\item{plot.pch=c(p.value=2,fdr=4,qval.0=3,e.value=1,fwer=20)}{Specific characters to distinguish the plotted statistics.}

\item{plot.col=c(p.value='#000000',fdr='#888888',qval.0='#666666',e.value='#BBBBBB',fwer='#444444')}{Specific colors or gray levels to distinguish the plotted statistics.}

\item{plot.elements=c("p.value","e.value","fwer","fdr","qval.0")}{Selection of elements to display on the plot.}

\item{plot.legend=TRUE}{Plot a legend indicating the number of features declared significant with the alpha threshold on the selected statistics.}

\item{legend.corner="topleft"}{corner wher the legend has to be placed.}

\item{legend.cex=1}{Font size for the legend.}

\item{xlab="p-value"}{}

\item{ylab="Multi-testing}{corrected statistics"}
}
\description{
Plot the different multi-testing corrected statistics as a
function of the nominal P-value.
}
\examples{
## To obtain the input list (multitest.result), run the example of
## stats4bioinfo::multipleTestingCorrections().
example(multipleTestingCorrections)

## Plot all the multiple testing corrections at once
mulitpleTestingCorrections.plotCorrectedVsPval(multitest.result)

## Compare e-value and FWER
mulitpleTestingCorrections.plotCorrectedVsPval(multitest.result, plot.elements=c("e.value","fwer"))

## Compare e-value and FDR.
## This plot highlights the non-linear relationship between FDR and p-value.
mulitpleTestingCorrections.plotCorrectedVsPval(multitest.result, plot.elements=c("e.value","fdr"))

## Compare Benjamini-Hochberg (qval.0) and Storey-Tibshirani (fdr) estimates of FDR.
mulitpleTestingCorrections.plotCorrectedVsPval(multitest.result, plot.elements=c("qval.0","fdr"))
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

