% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipleTestingCorrections.R
\name{mulitpleTestingCorrections.plotPvalDistrib}
\alias{mulitpleTestingCorrections.plotPvalDistrib}
\title{Multiple corrections for multiple testing}
\usage{
mulitpleTestingCorrections.plotPvalDistrib(multitest.result,
  main = "P-value distribution", plot.legend = TRUE, legend.cex = 1,
  legend.corner = "topright", breaks = seq(from = 0, to = 1, by = 0.05),
  draw.lambda = "arrow", draw.m0.line = TRUE, draw.mean.line = TRUE,
  overlay = NULL, col = "#FFEEDD", overlay.col = "#CCCCCC",
  mean.line.col = "darkred", m0.line.col = "blue", ...)
}
\arguments{
\item{multitest.result}{the list returned by the function multipleTestingCorrections().}

\item{...}{Additional parameters are passed to hist()}

\item{main='Multitesting}{corrections'  main title of the plot}

\item{plot.legend=TRUE}{Plot a legend with some indicative numbers (m0, m1, pi0).}

\item{legend.corner="topright"}{corner wher the legend has to be placed.}

\item{legend.cex=1}{Font size for the legend.}

\item{breaks=seq(from=0,to=1,}{by=0.05)    Breaks for the histogram}

\item{draw.lambda="arrow"}{Indicate the level of the lambda parameter. Supported
representations: "arrow", "line", "none".}

\item{draw.m0.line=TRUE}{Draw an horizontal line indicating the estimated m0
(number of trully null features) per bin.}

\item{draw.mean.line=FALSE}{Draw a dashed horizontal line indicating the mean
number of features per bin. The difference between this line and the "m0 per bin"
line reflects the importance of truly alternative features.}

\item{col="#BBBBFF"}{Histogram background color (passed to hist()).}

\item{overlay=NULL}{Boolean vector marking features of a special category
(e.g. truly null features). A colored histogram will be printed on top of the
main histogram, to indicate the number of features belonging to this group.}

\item{overlay.col="#CCCCCC"}{Color for the overlay histogram.}

\item{mean.line.col="black"}{Color to draw the line indicating the mean number of feature per bin.}

\item{m0.line.col="black"}{Color to draw the line indicating the estimated m0 per bin.}
}
\description{
Generate a plot that reproduces Fig 1 from Storey and Tibshirani (2003),
with some additional details, in order to illustrate the estimation of
the parameter PI0 = m0/m1.
}
\examples{
## To obtain the input list (multitest.result), run the examples of
## stats4bioinfo::multipleTestingCorrections().

example(multipleTestingCorrections)

## Plot the p-value distribution + landmarks
mulitpleTestingCorrections.plotPvalDistrib(multitest.result, draw.lambda="line")
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

