% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ROC.R
\name{plot.ROC}
\alias{plot.ROC}
\title{Plot a ROC curve}
\usage{
\method{plot}{ROC}(ROC.table, main = "ROC", xlab = "negative scores",
  ylab = "positive scores", add = FALSE, ...)
}
\arguments{
\item{ROC.table, }{## A data fame resulting from the function ROCstatistics()}

\item{...}{Additional parameters are passed to lines().}

\item{main='ROC'}{main title for the plot}

\item{xlab='negative}{scores'}

\item{ylab='positive}{scores'}

\item{add=FALSE}{If true, the ROC curve is added on a pre-existing plot}
}
\description{
Plot a ROC curve from a table returned by ROCstatistics.
}
\details{
First version: 2015-04
Last modification: 2015-04
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

