% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPvalCompa.R
\name{plotPvalCompa}
\alias{plotPvalCompa}
\title{Draw an XY plot comparing two series of p-values (or corrected p-values, or e-values).}
\usage{
plotPvalCompa(frame.to.plot, xlab = names(frame.to.plot)[1],
  ylab = names(frame.to.plot)[2], alpha = 0.05, score.name = "p-value",
  plot.result = TRUE, plot.colors = TRUE, plot.cex = 0.7,
  legend.cex = 0.9, legend.corner = "bottomright", ...)
}
\arguments{
\item{frame.to.plot}{A data.frame comprizing two columns, with the respective p-values (corrected or not) to plot.}

\item{...}{Additional parameters are passed to plot()}

\item{xlab=names(frame.to.plot)[1]}{Passed to plot()}

\item{ylab=names(frame.to.plot)[2]}{Passed to plot()}

\item{alpha=0.05}{Alpha threshold to select positives and compute the contingency table.}

\item{score.name="p-value"}{Score name to display on the legend (e.g. "p-value", "e-value", "fdr").}

\item{plot.result=TRUE}{Draw a comparison plot}

\item{plot.colors=TRUE}{Use colors for the plot}

\item{plot.cex=0.7}{Point size for the plot (passed to graphics::plot() function).}

\item{legend.cex=0.9}{Font size factor for legend. This allows to choose separate font sizes to draw the plot itself and the legend.}

\item{legend.corner="bottomright"}{Position where legend should be placed.}
}
\description{
Draw an XY plot comparing two series of
p-values (or corrected p-values, or e-values).
}
\details{
First version: 2015-03
Last modification: 2015-03
}
\examples{
################################################################
## Load dataset from DenBoer, 2009 and define two groups of interest
library(denboer2009)
group1 <- "Bo"
group2 <- "Bt"

verbose(paste("Selecting samples for groups", group1, "and", group2), 1)
selected.groups <- c(group1, group2)
selected.samples <- denboer2009.pheno$sample.labels \%in\% selected.groups

## Selected expression profiles
selected.probesets <- apply(denboer2009.amp == "P", 1, sum) >= 30
expr <- denboer2009.expr[selected.probesets, selected.samples]
sample.labels <- denboer2009.pheno$sample.labels[selected.samples]
verbose(paste("Expression profiles: ", nrow(expr), "probesets x", ncol(expr), "samples"), 1)
table(sample.labels)

################################################################
## Compare p-values returned by Student and Welch t-tests
student.result <- tTestPerRow(x=expr,cl = sample.labels, var.equal = TRUE)
welch.result <- tTestPerRow(x=expr,cl = sample.labels, var.equal = FALSE)
plotPvalCompa(data.frame("Student.e.value"=student.result$table$e.value,
                         "Welch.e.value"=welch.result$table$e.value),
              score.name = "e-value",
              legend.corner="bottomright")
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

