% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{plotROCframe}
\alias{plotROCframe}
\title{Plot an empty frame to draw a ROC curve}
\usage{
plotROCframe(main = "ROC curve", xlab = "FPR", ylab = "Sensitivity",
  xlim = c(0, 1), ylim = c(0, 1), grid.bars = seq(from = 0, to = 1, by =
  0.1), plot.limits = TRUE, plot.diagonal = TRUE, ...)
}
\arguments{
\item{...}{Additional parameters are passed to plot().}

\item{main="ROC}{curve"  Main title for the plot}

\item{xlab="False}{Positive Rate (FPR)"  Label for the abcsissa}

\item{ylab="True}{Positive Rate (TPR)" Label for the ordinate}

\item{xlim=c(0,1)}{Limits for the abcsissa}

\item{ylim=c(0,1)}{Limits for the ordinates}

\item{grid.bars=seq(from=0,to=1,by=0.1)}{Grid bars}

\item{plot.limits=TRUE}{Plot a black rectangle with the limits.}

\item{plot.diagonal=TRUE}{Plot the diagonal to indicate random expectation.}
}
\description{
Plot an empty frame to draw a ROC curve
}
\details{
First version: 2015-04
Last modification: 2015-04
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

