% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ROC_curves.R
\name{ROCstatistics}
\alias{ROCstatistics}
\title{ROC statistics}
\usage{
ROCstatistics(neg.scores, pos.scores, neg.total = length(neg.scores),
  pos.total = length(pos.scores), sort.decreasing = F)
}
\arguments{
\item{neg.scores}{}

\item{pos.scores}{}

\item{neg.total=length(neg.scores)}{}

\item{pos.total=length(pos.scores)}{}

\item{sort.decreasing=F}{Indicate whether the score should be sorted by
decreasing order (this option should be used when lower scores are
considered to be more significant, e.g. P-value, E-value, FDR).}
}
\description{
Compute the statistics required to draw a ROC curve,
+ some additional validation statistics
}
\details{
First version: 2015-04
Last modification: 2015-04
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

