% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tTestPerRow.bootstrap.R
\name{sampleClasses}
\alias{sampleClasses}
\title{Resample the elements of a named vector (typically class labels).}
\usage{
sampleClasses(cl, subset.sizes = NULL, replace = TRUE,
  discard.dup = FALSE)
}
\arguments{
\item{cl}{vector of class labels}

\item{subset.sizes=NULL}{named vector indicating the resampling size (values) for each class (names). If NULL, labels are resampled with the same frequency as in the input vector.}

\item{replace=TRUE}{}
}
\description{
Resample the elements of a named vector (typically class labels).
}
\details{
First version: 2015-04
Last modification: 2015-09
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

