% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqToPSSM.R
\name{seqToPSSM}
\alias{seqToPSSM}
\title{Build a position-specific scoring matrix (PSSM) from a set of sequences.}
\usage{
seqToPSSM(sequences, prior = NULL, pseudo.count = 2, IC.log.base = 2,
  case.sensitive = FALSE)
}
\arguments{
\item{sequences}{vector of strings corresponding to biological sequences (DNA, RNA, proteins)}

\item{prior=NULL}{vector of residue prior probabilities (names must correspond to residues)}

\item{pseudo.count=2}{pseudo-count}

\item{IC.log.base=2}{Logarithmic base for the information content}

\item{case.sensitive=FALSE}{by default residues are considered case-insensitive and converted to uppercases.}
}
\description{
Given a vector of sequences, built a position-specific scoring matrix (PSSM)
 with different derived statistics (counts, frequencies, probabilities, weights, 
information content).
}
\details{
First version: 2016-12-23
Last modification: 2016-12
}
\examples{
## Define the sequences of yeast Met31p binding sites
sequences <- c(
  "MET28"="cgcccAAAACTGTGGtgttag",
  "MET3"="gttgtAAAACTGTGGCTTTGT",
  "MUP3"="cggaaAAAACTGTGGcgtcgc",
  "SAM1"="acaggAAAACTGTGGtggcgc",
  "SAM2"="gcttgAAAACTGTGGcgtttt",
  "MET6"="gtcgcAAAACTGTGGtagtca",
  "MET30"="ccgcgCAAACTGTGGcttccc",
  "ZWF1"="ataagCAAACTGTGGgttcat",
  "MET14"="cctcaAAAAATGTGGcaatgg",
  "MET17"="tcatgAAAACTGTGTaacata",
  "MET2"="tgcaaAAAATTGTGGatgcac",
  "MET8"="ggaaaAAAAATGTGAaaatcg",
  "MET1"="cataaTAAACTGTGAacggac")

## Chose priors based on yeast non-coding sequences
prior <- c("A"=0.32, "C"=0.18, "G"=0.18, "T"=0.32)

## Build the PSSM
pssm <- seqToPSSM(seq=sequences, prior = prior)

## Print count table
print(pssm$counts)

## Print weight matrix
signif(pssm$weights, digits=2)

## Plot a heatmap with the weights
heatmap.simple(pssm$counts, auto.margins=FALSE, xlab="Position", 
     ylab="Residues", main="Yeast Met13p count matrix", las=1)

}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

