% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{shadeTails}
\alias{shadeTails}
\title{Plot a distribution and shade one or two tails.}
\usage{
shadeTails(fun, xlim, q, tails = "both", add = FALSE, ...)
}
\arguments{
\item{fun}{function}

\item{q}{quantile above (below) which the curve should be shaded on the upper (lower) tail.}

\item{...}{Additional parameters are passed to shadeArea()}

\item{tails="both"}{Tails under which the curve should be shaded. Accepted values: "both", "lower", "upper".}
}
\description{
Plot a distribution and shade one or two tails.
}
\details{
First version: 2015-03
Last modification: 2015-03
}
\examples{
## Shade the interval (-1,1) below Student density function
shadeTails(dt, xlim=c(-5,5), q=2,df=3, shade.col="pink", tails="both",
           curve.col="darkblue",shade.border="darkblue")
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

