% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats4bioinfo.R
\docType{package}
\name{stats4bioinfo}
\alias{stats4bioinfo}
\alias{stats4bioinfo-package}
\title{Statistics for bioinformatics - utilities and sample datasets}
\source{
\url{http://pedagogix-tagc.univ-mrs.fr/courses/statistics_bioinformatics/}
}
\description{
The package \code{stats4bioinfo} regroups various utilities implemented for the course "Statistics for bioinformatics" and the corresponding textbook, as well as some illustrative datasets.
}
\details{
\tabular{ll}{
     Package: \tab stats4bioinfo\cr
     Type: \tab Package\cr
     Version: \tab 1.0\cr
     Date: \tab 2015-03-03\cr
     License: \tab GNU\cr
  }
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}
\references{
Jacques van Helden (2016). Statistics for bioinformatics. Oxford University Press.
}
\seealso{
\code{\link{denboer2009}}
}

