% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statsPerRow.R
\name{statsPerRow}
\alias{statsPerRow}
\title{Descriptive statistics on each row of the input matrix}
\usage{
statsPerRow(x)
}
\arguments{
\item{x}{A matrix or data frame}
}
\value{
A data.frame with one row per row of the input matrix, and one column
per computed statistics.
\describe{
 \item{mean}{mean}
 \item{median}{median}
 \item{sd}{standard deviation}
 \item{var}{variance}
 \item{iqr}{inter-quartile range}
 \item{sdiqr}{standardized IQR, which can serve as robust estimator of the standard deviation.}
}
}
\description{
Compute descriptive parameters (central tendency, dispersion)
for each row of a matrix or data frame.
}
\details{
First version: 2015-03
Last modification: 2015-03
}
\examples{
## Load example data set from Den Boer, 2009
library(denboer2009)
data(denboer2009.expr)     ## Load expression table
data(denboer2009.pheno)    ## Load phenotypic data
data(denboer2009.group.labels)    ## Load phenotypic data

stats.per.row <- statsPerRow(denboer2009.expr)
names(stats.per.row)
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

