% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tTestPerRow.R
\name{stderPlot}
\alias{stderPlot}
\title{Standard error plot}
\usage{
stderPlot(multi.t, col.points = "#888888", col.grid = "#CCCCCC",
  col.lines = "blue", col.positive = "#00BB00", ...)
}
\arguments{
\item{multi.t}{Must be the data frame obtained with tTestPerRow()}

\item{...}{Additional parameters are passed to the plot function}

\item{col.points='#888888'}{Color(s) for the points}

\item{col.lines='blue'}{Color for the line highlighting the significance threshold}

\item{col.grid=#CCCCCC}{Grid color}

\item{col.positive='#00BB00'}{Color to highlight significant points (called positive)}
}
\description{
Complement to the volcano plots generated by stats4bioinfo::tTestPerRow.plotVolcano()
Draw a plot that enables to compare
- standard error and significance
- mean difference, standard error and significance
}
\examples{
## The object welch.table should have been generated with \\code{\\link{tTestPerRow}} (see examples in the doc)
stderPlot(welch.table)
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

