% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tTestPerRow.bootstrap.R
\name{tTestPerRow.bootstrap.hist}
\alias{tTestPerRow.bootstrap.hist}
\title{Plot an histogram with a bootstrap test result.}
\usage{
tTestPerRow.bootstrap.hist(bootstrap.result, xlab = "Support",
  ylab = "Number of features", support.statistics = "fdr",
  support.quantile = bootstrap.result$support.quantile, main = paste(sep =
  "", "Bootstrap support (", support.statistics, "<=", bootstrap.result$alpha,
  ")"), plot.dcdf = FALSE, plot.legend = TRUE, legend.corner = "topright",
  lwd = 1, ylim = NULL, ...)
}
\arguments{
\item{bootstrap.result}{An object returned by the function tTestPerRow.bootstrap().}

\item{plot.legend}{= TRUE  Only valid if plot.dcdf is TRUE. Plot a legend indicating
the number of features supported by a given number of the iterations.}

\item{...}{Additional parameters are passed to hist().}

\item{support.statistics="fdr"}{Statistics for which the support has to be displayed.
Supported values: "fdr", "p.value", "e.value".}

\item{support.quantile=0.75}{Minimal percent of support required to declare a feature positive.
The default is to select features supported in 75\% of bootstrap iterations.}

\item{xlab="Support"}{}

\item{ylab="Number}{of features"}

\item{main=paste(sep="",}{"Bootstrap support (", support.statistics,"<=", bootstrap.result$alpha, ")")}

\item{plot.dcdf=FALSE}{Plot the decreasing CDF over the histogram, in "step" mode.
This curve indicates the number of features supported by at least X bootstraps.}

\item{legend.corner="topright"}{Corner to plo the legend.}

\item{lwd=1}{Line width for the dcdf and support threshold lines.}
}
\description{
Plot an histogram from the result of tTestPerRow.bootstrap().
The histogram indicates the number of features (ordinate) with a given support
(abcissa), i.e. the number of features declared significant in exactly
X bootstrap iterations.
}
\details{
First version: 2015-04
Last modification: 2015-04
}
\examples{
## First run the examples of tTestPerRow.bootstrap(), in order to get some result to plot
example("tTestPerRow.bootstrap")

## Plot the histogram
tTestPerRow.bootstrap.hist(student.bootstrap, col="#BBBBBB")

## Plot the histogram + the dcdf.
tTestPerRow.bootstrap.hist(student.bootstrap, col="#BBBBBB",
                           plot.dcdf=TRUE, lwd=2)

## Plot the histogram + the dcdf for the e-value.
tTestPerRow.bootstrap.hist(student.bootstrap, support.statistics="e.value",
                           col="#BBBBBB",
                           plot.dcdf=TRUE, lwd=2)
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

