% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tTestPerRow.plotVolcano.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Draw a Volcano plot.}
\usage{
VolcanoPlot(multitest.table, effect.size.col, control.type = "p.value",
  alpha = 0.05, effect.threshold = NULL, sort.by.pval = FALSE,
  plot.points = TRUE, plot.ci = FALSE, xlab = "Effect size",
  ylab = paste(sep = "", "-log10(", control.type, ")"),
  col.points = "#888888", col.positive = "#008800", col.grid = "#BBBBBB",
  col.lines = "blue", pch.positive = 3, pch.negative = 1, xlim = NULL,
  ylim = NULL, cex = 0.6, legend.corner = "bottomleft",
  full.legend = FALSE, legend.cex = 1, tick.size = 0.05, ...)
}
\arguments{
\item{effect.size.col}{A column number or name, inidicating which column of the result table contains the effect size, which will be}

\item{xlim}{Range of the X axis.}

\item{ylim}{Range of the Y axis.}

\item{...}{Additional parameters are passed to plot()}

\item{result.table}{A data frame containing one row per feature, and one column per statistics.}

\item{control.type="p.value"}{A column number or name, indicating which column of the result table contains the p-value or an equivalent indication of the significance of each feature (example: "fdr", "e.value", "p.value")}

\item{alpha=0.05}{Alpha threshold for the control of false positives}

\item{effect.threshold=NULL}{Threshold on the absolute value of the effect size.}

\item{sort.by.pval=FALSE}{Sort row by p-value in order to plot significant elements on top of non-significant}

\item{xlab="Effect}{size" Label for the X axis.}

\item{ylab="sig}{= -log10(p-value)" Label for the Y axis}

\item{col.points='#888888'}{Color(s) for the points. can be either a single value (same color for all points), or a vector of the same length as the numer of genes (rows) of the input table.}

\item{col.positive='#008800'}{Color to highlight significant points (called positive). When NULL, positive points are not displayed.}

\item{col.lines='blue'}{Color for the line highlighting the significance threshold}

\item{col.grid=#CCCCCC}{Grid color}

\item{pch.positive=3}{Point shape for the positive tests (default: 3, i.e. + symbol)}

\item{pch.negative=1}{Point shape for the negative tests (default: 1, i.e. o symbol)}

\item{legend.corner="bottomleft"}{Corner for the legend. When NULL, no legend is displayed.}

\item{full.legend=TRUE}{Plot additional indications on the legend (number of elements passing the different adjusted p-values)}

\item{legend.cex=1}{Font size for the legend.}

\item{cex=0.6}{Point size, passed to plot() and lines()}

\item{plot.points=TRUE}{Plot one point per row with the significance (sig=-log10(p-value)) as a function of the effect size.}

\item{plot.ci=FALSE}{Plot horizontal bars to denote confidence intervals around the mean difference.}

\item{tick.size=0.05}{Height of the vertical lines denoting the boundaries of confidence intervals.}
}
\value{
no return object
}
\description{
Draw a volcano plot from a table containing at least one column for the effect size, and another one for the p-value or an equivalent measure of significance (FDR, e-value, FWER, ...).
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

