% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCstatistics.R
\name{ROCstatistics}
\alias{ROCstatistics}
\title{ROC statistics}
\usage{
ROCstatistics(neg.scores, pos.scores, neg.total = length(neg.scores),
  pos.total = length(pos.scores), sort.decreasing = FALSE)
}
\arguments{
\item{neg.scores}{vector of scores obtained with the truly negative cases}

\item{pos.scores}{vector of scores obtained with the truly positive cases}

\item{neg.total=length(neg.scores)}{total number of negative cases}

\item{pos.total=length(pos.scores)}{total number of positive cases}

\item{sort.decreasing=FALSE}{Indicate whether the score should be sorted by
decreasing order (this option should be used when lower scores are 
considered to be more significant, e.g. P-value, E-value, FDR).}
}
\description{
Compute the statistics required to draw a ROC curve, 
+ some additional validation statistics
}
\details{
First version: 2015-04
Last modification: 2015-04
}
\author{
Jacques van Helden (\email{Jacques.van-Helden@univ-amu.fr})
}

